feat(dashboard): edit variable popover autosave
