+++ apps/dashboard/src/components/primitives/control-input/control-input.tsx
@@ -66,7 +66,7 @@ export function ControlInput({
     viewRef,
     onChange
   );
-
+  const isVariablePopoverOpen = !!selectedVariable;
   const variable: LiquidVariable | undefined = selectedVariable
     ? {
         name: selectedVariable.value,
@@ -109,6 +109,7 @@ export function ControlInput({
     (open: boolean) => {
       if (!open) {
         setTimeout(() => setSelectedVariable(null), 0);
+        viewRef.current?.focus();
       }
     },
     [setSelectedVariable]
@@ -129,19 +130,21 @@ export function ControlInput({
         value={value}
         onChange={onChange}
       />
-      <EditVariablePopover
-        open={!!selectedVariable}
-        onOpenChange={handleOpenChange}
-        variable={variable}
-        isAllowedVariable={isAllowedVariable}
-        onUpdate={(newValue) => {
-          handleVariableUpdate(newValue);
-          // Focus back to the editor after updating the variable
-          viewRef.current?.focus();
-        }}
-      >
-        <div />
-      </EditVariablePopover>
+      {isVariablePopoverOpen && (
+        <EditVariablePopover
+          open={isVariablePopoverOpen}
+          onOpenChange={handleOpenChange}
+          variable={variable}
+          isAllowedVariable={isAllowedVariable}
+          onUpdate={(newValue) => {
+            handleVariableUpdate(newValue);
+            // Focus back to the editor after updating the variable
+            setTimeout(() => viewRef.current?.focus(), 0);
+          }}
+        >
+          <div />
+        </EditVariablePopover>
+      )}
     </div>
   );
 }
