+++ apps/dashboard/src/components/variable/types.ts
@@ -22,6 +22,5 @@ export type FilterWithParam = {
 
 export type VariablePopoverProps = {
   variable?: LiquidVariable;
-  onUpdate: (newValue: string) => void;
   onEscapeKeyDown?: (event: KeyboardEvent) => void;
 };
