+++ apps/dashboard/src/components/workflow-editor/steps/step-drawer.tsx
@@ -1,12 +1,15 @@
+import { useCallback, useId } from 'react';
 import { motion } from 'motion/react';
 import { useNavigate } from 'react-router-dom';
+import { StepTypeEnum } from '@novu/shared';
 
 import { Sheet, SheetContentBase, SheetDescription, SheetPortal, SheetTitle } from '@/components/primitives/sheet';
 import { VisuallyHidden } from '@/components/primitives/visually-hidden';
 import { PageMeta } from '@/components/page-meta';
 import { useWorkflow } from '@/components/workflow-editor/workflow-provider';
-import { StepTypeEnum } from '@novu/shared';
 import { cn } from '@/utils/ui';
+import { useEscapeKeyManager } from '@/context/escape-key-manager/hooks';
+import { EscapeKeyManagerPriority } from '@/context/escape-key-manager/priority';
 
 const transitionSetting = { ease: [0.29, 0.83, 0.57, 0.99], duration: 0.4 };
 const stepTypeToClassname: Record<string, string | undefined> = {
@@ -15,15 +18,18 @@ const stepTypeToClassname: Record<string, string | undefined> = {
 };
 
 export const StepDrawer = ({ children, title }: { children: React.ReactNode; title?: string }) => {
+  const id = useId();
   const navigate = useNavigate();
   const { workflow, step } = useWorkflow();
 
-  const handleCloseSheet = () => {
+  const handleCloseSheet = useCallback(() => {
     if (step) {
       // Do not use relative path here, calling twice will result in moving further back
       navigate(`../steps/${step.slug}`);
     }
-  };
+  }, [navigate, step]);
+
+  useEscapeKeyManager(id, handleCloseSheet, EscapeKeyManagerPriority.SHEET);
 
   if (!workflow || !step) {
     return null;
@@ -57,7 +63,6 @@ export const StepDrawer = ({ children, title }: { children: React.ReactNode; tit
               // for example maily variable list, the conditions operators
               e.preventDefault();
             }}
-            onEscapeKeyDown={handleCloseSheet}
           >
             <motion.div
               initial={{
