+++ apps/dashboard/src/context/escape-key-manager/escape-key-context.tsx
@@ -1,0 +1,12 @@
+import { createContext } from 'react';
+import { EscapeKeyManagerPriority } from './priority';
+
+export type EscapeKeyManagerContextType = {
+  registerEscapeHandler: (id: string, handler: () => void, priority?: EscapeKeyManagerPriority) => void;
+  unregisterEscapeHandler: (id: string) => void;
+};
+
+export const EscapeKeyManagerContext = createContext<EscapeKeyManagerContextType>({
+  registerEscapeHandler: () => {},
+  unregisterEscapeHandler: () => {},
+});
