+++ apps/dashboard/src/context/escape-key-manager/priority.ts
@@ -1,0 +1,5 @@
+export enum EscapeKeyManagerPriority {
+  NONE = 0,
+  SHEET = 100,
+  POPOVER = 200,
+}
